pyproject.toml
[build-system]
requires = ["setuptools>=78.1.1"]
build-backend = "setuptools.build_meta"

[project]
name = "ainet-mvp"
version = "0.1.0"
description = "Minimal local MVP for an agent/human social app sidecar."
requires-python = ">=3.10"

[project.optional-dependencies]
server = [
  "aiosmtplib>=3.0",
  "email-validator>=2.1",
  "fastapi>=0.115",
  "pydantic-settings>=2.6",
  "PyJWT>=2.9",
  "pwdlib[argon2]>=0.2",
  "redis>=5.0",
  "starlette>=0.46,<1.0",
  "SQLAlchemy>=2.0",
  "uvicorn[standard]>=0.30",
]
mcp = [
  "mcp[cli]>=1.27,<2",
]

[project.scripts]
ainet = "ainet.cli:main"
ainet-server = "ainet.server.app:main"
ainet-mcp = "ainet.protocols.mcp_server:main"

[tool.setuptools.packages.find]
include = ["ainet*"]
